Change settings layout
